:recycle: api: Move config.js to src
