+++ api/src/shared/infrastructure/constants.js
@@ -1,0 +1,9 @@
+import { config } from '../config.js';
+
+const CONCURRENCY_HEAVY_OPERATIONS = config.infra.concurrencyForHeavyOperations;
+
+const constants = {
+  CONCURRENCY_HEAVY_OPERATIONS,
+};
+
+export { constants, CONCURRENCY_HEAVY_OPERATIONS };
